Fix program termination occurring when min_samples was not low enough
This commit fixes a program termination that was occurring every time a capture was including local MAC addresses all with a number of samples lower than min_samples. Now the exception is caught, and these local MAC addresses will simply not be counted, returning a total count equal to the number of global MAC addresses, instead of just terminating the execution.